fix(category): migrate id type from Byte to Long

+++ src/main/java/dev/tchiwara/ecommerce/api/category/CategoryRepository.java
@@ -2,5 +2,5 @@ package dev.tchiwara.ecommerce.api.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category,Byte> {
+public interface CategoryRepository extends JpaRepository<Category,Long> {
 }
