+++ src/main/java/dev/tchiwara/ecommerce/api/category/dtos/CategoryResponseDTO.java
@@ -6,6 +6,6 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class CategoryResponseDTO {
-    private final Byte id;
+    private final Long id;
     private final String name;
 }
